requirements.txt
pandas==2.1.4
gspread==5.12.4
oauth2client==4.1.3
schedule==1.2.1
python-dotenv==1.0.0